Move test data generation to a scripts
