+++ Makefile.pre.in
@@ -958,6 +958,11 @@ regen-test-frozenmain: $(BUILDPYTHON)
 	# using Programs/freeze_test_frozenmain.py
 	$(RUNSHARED) ./$(BUILDPYTHON) $(srcdir)/Programs/freeze_test_frozenmain.py Programs/test_frozenmain.h
 
+.PHONY: regen-test-levenshtein
+regen-test-levenshtein: $(BUILDPYTHON)
+	# Regenerate Lib/test/levenshtein_examples.json
+	$(RUNSHARED) ./$(BUILDPYTHON) $(srcdir)/Tools/scripts/generate_levenshtein_examples.py Lib/test/levenshtein_examples.json
+
 .PHONY: regen-re
 regen-re: $(BUILDPYTHON)
 	# Regenerate Lib/re/_casefix.py
@@ -1223,7 +1228,7 @@ regen-limited-abi: all
 regen-all: regen-opcode regen-opcode-targets regen-typeslots \
 	regen-token regen-ast regen-keyword regen-sre regen-frozen clinic \
 	regen-pegen-metaparser regen-pegen regen-test-frozenmain \
-	regen-global-objects
+	regen-test-levenshtein regen-global-objects
 	@echo
 	@echo "Note: make regen-stdlib-module-names and make regen-configure should be run manually"
